requirements.txt
easyocr
langchain
openai
tiktoken
duckduckgo-search
youtube-transcript-api
pytube
textbase-client
poetry
requests